vertex_api/requirements-dev.txt
decorator==4.1.2
Django==1.11.4
django-debug-toolbar==1.8
django-extensions==1.9.0
django-reversion==2.0.10
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
rules==1.2.1
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
wcwidth==0.1.7
django-material==1.0.0
django-colorful==1.2
graphene-django>=2.0
